refactor(transfer-detection): honest skip log; drop dead SyncedAccountStore.transactions; share test double (Task 11 review)
