+++ MoolahBenchmarks/CryptoSyncBenchmarks.swift
@@ -188,9 +188,8 @@ final class CryptoSyncBenchmarks: XCTestCase {
       accounts: backend.accounts,
       transferDetection: TransferDetectionCoordinator(
         transactions: backend.transactions,
-        dismissedPairs: backend.dismissedTransferPairs,
+        suggestions: backend.transferSuggestions,
         clock: { Self.pinnedNow }),
-      transactions: backend.transactions,
       clock: { Self.pinnedNow },
       staleThreshold: 0)
   }
