+++ MoolahTests/Features/Crypto/CryptoAccountCreationStoreTests.swift
@@ -64,9 +64,8 @@ struct CryptoAccountCreationStoreTests {
       accounts: backend.accounts,
       transferDetection: TransferDetectionCoordinator(
         transactions: backend.transactions,
-        dismissedPairs: backend.dismissedTransferPairs,
+        suggestions: backend.transferSuggestions,
         clock: { Self.pinnedNow }),
-      transactions: backend.transactions,
       clock: { Self.pinnedNow })
     let accountStore = AccountStore(
       repository: backend.accounts,
