+++ MoolahTests/Features/Crypto/SyncedAccountStorePipelineStructureTests.swift
@@ -68,9 +68,8 @@ struct SyncedAccountStorePipelineStructureTests {
       accounts: backend.accounts,
       transferDetection: TransferDetectionCoordinator(
         transactions: backend.transactions,
-        dismissedPairs: backend.dismissedTransferPairs,
+        suggestions: backend.transferSuggestions,
         clock: { Self.pinnedNow }),
-      transactions: backend.transactions,
       clock: { Self.pinnedNow },
       staleThreshold: 86_400,
       timerInterval: .seconds(3_600),
