+++ MoolahTests/Features/Sync/SyncedAccountStoreGlobalErrorTests.swift
@@ -74,9 +74,8 @@ struct SyncedAccountStoreGlobalErrorTests {
       accounts: backend.accounts,
       transferDetection: TransferDetectionCoordinator(
         transactions: backend.transactions,
-        dismissedPairs: backend.dismissedTransferPairs,
+        suggestions: backend.transferSuggestions,
         clock: { Self.pinnedNow }),
-      transactions: backend.transactions,
       clock: { Self.pinnedNow })
     return Fixture(
       store: store, backend: backend, database: database, alchemy: alchemy)
